Replace finite automation with LL(1) recursive descent parser...
At the same time, we also replace the input grammar with a new (and simpler) one. We do this because in VS Code 1.38, the negative prefix for numbers (to calculate line numbers from the bottom of the document) was introduced for the 'Go To Line...' feature. In order to maintain parity with the built in feature, we replace the previous 'relative number' mode with the new behavior.

Furthermore, the new grammar changes the range notation from `:` (select) and `;` (quick select) to `..` and`.` respectively. This allows us to restore `:` for use as the separator between line and column numbers.