Update the grammar spec to have a negative number token...
Prior to this we were using two tokens to specify a line number relative to the bottom of the document. This meant that we would've allowed whitespace in between the `-` sign and the number that follows. We don't want this behavior, so we merge the `-` sign and the number that follows into one token.